Clean up + simplify lbAudio (#3217)
* Clean up + simplify lbAudio

* Simplify functions with inlines

lbAudio has quite a few internal functions which call each other, so by
using auto-inlining whenever possible we can greatly simplify the code.

* Maximally privatize member functions + data

The surface area of lbAudio is quite small, and is an opaque wrapper
over the underlying HSD + Dolphin audio libs. By privatizing anything
not used directly by the melee code, we can better understand the API
that this library actually exposes.

* Name well-known arguments and private functions/data

Renaming the public functions has potential to cause conflicts, so
currently deferring to a later PR. But most of them are well-known by
now, and we can give more descriptive names to most internal details.
The naming here was all done by propagating existing known names from
other function parameters, struct members and other parts of the code.

* Other misc cleanup
  * Use LANG enum values
  * Prefer standard int types, preserve u32/u64 for bitfields
  * Remove all gotos
  * Remove all unsafe data offsets

+++ src/sysdolphin/baselib/axdriver.h
@@ -67,7 +67,7 @@ struct HSD_SM {
                                   void* param);
 /* 38E034 */ s32 HSD_AudioGetAuxHeapSize(AXDriverAuxType type, void* param);
 /* 38E30C */ bool AXDriver_8038E30C(s32 channel, s32 type, void* param,
-                                    u8* heap, u32 heap_size);
+                                    u8* heap, size_t heap_size);
 /* 38E37C */ bool AXDriver_8038E37C(AXDriverAuxType type, void* param);
 /* 38E498 */ void AXDriver_8038E498(int voices, int priority, int sample_rate,
                                     int aram_size);
